Moved functions to UserService
